% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_calc_metrics.R
\name{calc_metrics}
\alias{calc_metrics}
\title{Calculate Composite Score Metrics}
\usage{
calc_metrics(df, composite_score, weights, digits, name)
}
\arguments{
\item{df}{A dataframe object. This should be a structured dataset where each
column represents a variable and each row represents an observation.}

\item{composite_score}{An array representing the composite score.}

\item{weights}{A vector of weights for the indicator variables.}

\item{digits}{The decimal places for the metrics to be rounded to. Default is
3.}

\item{name}{A string representing the name of the composite score.}
}
\value{
A list of dataframes consisting of:
\itemize{
 \item \code{composite_score}: An array with the calculated composite
 variable.
 \item \code{composite_metrics}: A matrix loadings and weights
 of the indicators.
 \item \code{composite_validity}: A matrix of composite reliability and
 validity metrics.
}
}
\description{
Calculate the indicator and composite-level reliability and
  validity metrics.
}
\keyword{internal}
