% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mi_composite.R
\name{calc_mi_composite}
\alias{calc_mi_composite}
\title{Calculate Mutual-Information Composite Scores}
\usage{
calc_mi_composite(
  data,
  var,
  entropy = "emp",
  digits = 3,
  name = NULL,
  return_metrics
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{var}{A required vector of indicator column names.}

\item{entropy}{A string value reflecting the mutual information entropy
estimator from the \code{infotheo} package. Four estimators are available:
\code{emp} to compute the entropy of the empirical probability
distribution. Empirical entropy is suitable for simple calculations without
corrections. \code{mm} applies an asymptotic bias-corrected estimator
making it suitable for small sample sizes. \code{shrink} applies a
shrinkage estimate of the Dirichlet probability distribution to provide a
stable estimate useful for small sample sizes or sparse data. \code{sg}
applies a Schurmann-Grassberger estimate of the Dirichlet probability
distribution to serve as an alternative to the Shrinkage approach.}

\item{digits}{The decimal places for the metrics to be rounded to. Default is
3.}

\item{name}{A required string denoting the name of the composite variable.}

\item{return_metrics}{Logic to determine whether to return reliability and
validity metrics. Set to \code{TRUE} for a list of dataframes with
reliability and validity metrics.}
}
\value{
If \code{return_metrics = FALSE}, an array of the composite score is
  returned. If \code{return_metrics = TRUE}, a list is returned consisting
  of:
\itemize{
 \item{\code{composite_score}: }{An array with the calculated composite
 variable.} \item{\code{composite_metrics}: }{A matrix loadings and weights
 of the indicators.}
 \item{\code{composite_validity}: }{A matrix of composite reliability and 
 validity metrics.}
}
}
\description{
Calculate the composite score for the mutual information family of weighting schemas.
For information on the specifics calculations, refer to the help
documentations of \code{?average_score} (for unweighted),
\code{?correlation_score} (for correlation-weighted), and
\code{?regression_score} (for regression-weighted).

Refer to help documentation \code{?calc_metrics} for information on how
reliability and validity metrics are calculated.
}
\examples{

data(grit)

# Specify a vector of indicators
extraversion <- sprintf("e\%01d", seq(10))

# Calculate mutual information composite score
calc_mi_composite(data = grit,
                  var = extraversion,
                  entropy = "emp",
                  weight = "correlation",
                  return_metrics = FALSE)

# Calculate mutual information composite score and metrics
calc_mi_composite(data = grit,
                  var = extraversion,
                  entropy = "emp",
                  name = "extraversion",
                  digits = 3,
                  return_metrics = TRUE)

}
