% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sd_composite.R
\name{calc_sd_composite}
\alias{calc_sd_composite}
\title{Calculate Standard Deviation Composite Scores}
\usage{
calc_sd_composite(data, var, weight, digits = 3, name = NULL, return_metrics)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{var}{A required vector of indicator column names.}

\item{weight}{Required weighting schema. Schemas include
\code{c("upweight", "downweight")}}

\item{digits}{The decimal places for the metrics to be rounded to. Default is
3.}

\item{name}{A required string denoting the name of the composite variable.}

\item{return_metrics}{Logic to determine whether to return reliability and
validity metrics. Set to \code{TRUE} for a list of dataframes with
reliability and validity metrics.}
}
\value{
If \code{return_metrics = FALSE}, an array of the composite score is
  returned. If \code{return_metrics = TRUE}, a list is returned consisting
  of:
\itemize{
 \item{\code{composite_score}: }{An array with the calculated composite
 variable.} \item{\code{composite_metrics}: }{A matrix loadings and weights
 of the indicators.}
 \item{\code{composite_validity}: }{A matrix of composite reliability and 
 validity metrics.}
}
}
\description{
Calculate the composite score for the standard deviation family of weighting
schemas. For information on the specifics calculations, refer to the help
documentations of \code{?sd_upweight_score} (for upweighted standard
deviation) and \code{?sd_downweight_score} (for downweighted standard
deviation).

Refer to help documentation \code{?calc_metrics} for information on how
reliability and validity metrics are calculated.
}
\examples{

data(grit)

# Specify a vector of indicators
extraversion <- sprintf("e\%01d", seq(10))

# Calculate correlation-weighted composite score
calc_sd_composite(data = grit,
                  var = extraversion,
                  weight = "sd_upweight",
                  return_metrics = FALSE)

# Calculate correlation-weighted composite score and metrics
calc_sd_composite(data = grit,
                  var = extraversion,
                  weight = "sd_upweight",
                  name = "extraversion",
                  digits = 3,
                  return_metrics = TRUE)

}
