% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_calc_single_indicator_composite.R
\name{calc_single_indicator}
\alias{calc_single_indicator}
\title{Calculate Single Indicator Composite}
\usage{
calc_single_indicator(data, var, name, digits = 3, return_metrics = FALSE)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{var}{A required vector of indicator column name.}

\item{name}{A required string denoting the name of the composite variable.}

\item{digits}{The decimal places for the metrics to be rounded to. Default is
3. This argument is only relevant if \code{return_metrics = TRUE}.}

\item{return_metrics}{Logic to determine whether to return reliability and
validity metrics. Set to \code{TRUE} for a list of dataframes with
reliability and validity metrics.}
}
\description{
Return the indicator as its own composite.
}
\keyword{internal}
