% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_detect_higher_order_variable.R
\name{detect_higher_order_variable}
\alias{detect_higher_order_variable}
\title{Detect Higher Order Constructs}
\usage{
detect_higher_order_variable(composite_list)
}
\arguments{
\item{composite_list}{A required named list of vectors. Each name in the list
represents a composite variable, and the corresponding vector contains the
column names that are associated with the indicators comprising said
composite variable.}
}
\value{
A vector of higher order construct/variable names. The function uses
  a for loop to iterate over each named vector in the list and checks if any
  of the values in the vector match the names of the vectors in the list. If
  a match is found, the name of that vector is added to the
  \code{higher_order_variables} vector. The function then returns
  \code{higher_order_variables.}
}
\description{
Takes a composite list as input and returns a character vector
  containing the names of all composite variable vectors that are considered
  'higher order' constructs/variables. A construct is considered at the
  'higher order' if it is a composite of other composite scores that
  represent different dimensions. This function automatically detects higher
  order constructs as any vector containing one or more of the names of other
  composite vectors in the list.
}
\seealso{
[remove_higher_order_variable(), keep_higher_order_variable()]
}
\keyword{internal}
