% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grit.R
\docType{data}
\name{grit}
\alias{grit}
\title{Grit Data}
\format{
An object of class \code{data.frame}. The items are listed below.
  Items with asterisks (*) are reverse coded. All items are already reverse
  coded in the dataset for convenience:
\enumerate{
 \item \code{id} Respondent ID
 \item Extraversion
 \enumerate{
 \item \code{e1} I am the life of the party.
 \item \code{e2} I don't talk a lot.*
 \item \code{e3} I feel comfortable around people.
 \item \code{e4} I keep in the background.*
 \item \code{e5} I start conversations.
 \item \code{e6} I have little to say.*
 \item \code{e7} I talk to a lot of different people at parties.
 \item \code{e8} I don't like to draw attention to myself.*
 \item \code{e9} I don't mind being the center of attention.
 \item \code{e10} I am quiet around strangers.*
 }
 \item Neuroticism
 \enumerate{
 \item \code{n1} I get stressed out easily.
 \item \code{n2} I am relaxed most of the time.*
 \item \code{n3} I worry about things.
 \item \code{n4} I seldom feel blue.*
 \item \code{n5} I am easily disturbed.
 \item \code{n6} I get upset easily.
 \item \code{n7} I change my mood a lot.
 \item \code{n8} I have frequent mood swings.
 \item \code{n9} I get irritated easily.
 \item \code{n10} I often feel blue.
 }
 \item Agreeableness
 \enumerate{
 \item \code{a1} I feel little concern for others.*
 \item \code{a2} I am interested in people.
 \item \code{a3} I insult people.*
 \item \code{a4} I sympathize with others' feelings.
 \item \code{a5} I am not interested in other people's problems.*
 \item \code{a6} I have a soft heart.
 \item \code{a7} I am not really interested in others.*
 \item \code{a8} I take time out for others.
 \item \code{a9} I feel others' emotions.
 \item \code{a10} I make people feel at ease.
 }
 \item Conscientiousness
 \enumerate{
 \item \code{c1} I am always prepared.
 \item \code{c2} I leave my belongings around.*
 \item \code{c3} I pay attention to details.
 \item \code{c4} I make a mess of things.*
 \item \code{c5} I get chores done right away.
 \item \code{c6} I often forget to put things back in their proper place.*
 \item \code{c7} I like order.
 \item \code{c8} I shirk my duties.*
 \item \code{c9} I follow a schedule.
 \item \code{c10} I am exacting in my work.
 }
 \item Openness
 \enumerate{
 \item \code{o1} I have a rich vocabulary.
 \item \code{o2} I have difficulty understanding abstract ideas.*
 \item \code{o3} I have a vivid imagination.
 \item \code{o4} I am not interested in abstract ideas.*
 \item \code{o5} I have excellent ideas.
 \item \code{o6} I do not have a good imagination.*
 \item \code{o7} I am quick to understand things.
 \item \code{o8} I use difficult words.
 \item \code{o9} I spend time reflecting on things.
 \item \code{o10} I am full of ideas.
 }
 \item Grit
 \enumerate{
 \item \code{gs1} I have overcome setbacks to conquer an important challenge
 \item \code{gs2} New ideas and projects sometimes distract me from previous ones.*
 \item \code{gs3} My interests change from year to year.*
 \item \code{gs4} Setbacks don’t discourage me.
 \item \code{gs5} I have been obsessed with a certain idea or project for a short time but later lost interest.*
 \item \code{gs6} I am a hard worker.
 \item \code{gs7} I often set a goal but later choose to pursue a different one.*
 \item \code{gs8} I have difficulty maintaining my focus on projects that take more than a few months to complete.*
 \item \code{gs9} I finish whatever I begin.
 \item \code{gs10} I have achieved a goal that took years of work.
 \item \code{gs11} I become interested in new pursuits every few months.*
 \item \code{gs12} I am diligent.
 }
}
}
\source{
\href{https://openpsychometrics.org/_rawdata/}{OpenPsychoMetrics}.
}
\usage{
data(grit)
}
\description{
This example data contains 500 responses from people completing
  the Big 5 personality inventory (John & Srivastava, 1999) and the Grit
  Scale (Duckworth et al., 2007) from
  \href{https://openpsychometrics.org/_rawdata/}{OpenPsychoMetrics}.
}
\references{
Duckworth, A.L., Peterson, C., Matthews, M.D., & Kelly, D.R.
  (2007). Grit: Perseverance and passion for long-term goals. Journal of
  Personality and Social Psychology, 9, 1087-1101.

  John, O. P., & Srivastava, S. (1999). The Big-Five trait taxonomy: History,
  measurement, and theoretical perspectives. In L. A. Pervin & O. P. John
  (Eds.), Handbook of personality: Theory and research (2nd ed., pp.
  102–138). Guilford Press.
}
\keyword{datasets}
