% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\title{Create Link Paths Between Node Sets}
\usage{
link(from, to)
}
\arguments{
\item{from}{A character vector of source node names.}

\item{to}{A character vector of target node names.}
}
\value{
A data frame with two columns, `from` and `to`, containing all
  directed edges from each element in `from` to each element in `to`.
}
\description{
Generates a set of directed paths representing all combinations of links from
nodes in `from` to nodes in `to`. The result is a data frame that can be used
as input to `composite_model()` and related functions.
}
\examples{
link(from = c("x1", "x2"), to = c("y"))

}
