% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_calc_median_composite.R
\name{calc_median_composite}
\alias{calc_median_composite}
\title{Calculate Median Composite Scores}
\usage{
calc_median_composite(
  data,
  var,
  weight,
  digits = 3,
  name = NULL,
  decay_rate = 0.5,
  sigma = 0.5,
  return_metrics
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{var}{A required vector of indicator column names.}

\item{weight}{Required weighting schema. Schemas include \code{c("median",
"median_decay", "median_gauss")}}

\item{digits}{The decimal places for the metrics to be rounded to. Default is
3.}

\item{name}{A required string denoting the name of the composite variable.}

\item{decay_rate}{A numeric value reflecting the decay rate (i.e.,
sensitivity) of the distance-to-median weighting schema. The default value
is set to 0.5. This argument is only relevant if \code{weight =
"median_decay"}.}

\item{sigma}{A numeric value reflecting the sigma value for the Gaussian
function in the distance-to-median weighting schema. The default value is
set to 0.5. This argument is only relevant if \code{weight =
"median_gauss"}.}

\item{return_metrics}{Logic to determine whether to return reliability and
validity metrics. Set to \code{TRUE} for a list of dataframes with
reliability and validity metrics.}
}
\value{
If \code{return_metrics = FALSE}, an array of the composite score is
  returned. If \code{return_metrics = TRUE}, a list is returned consisting
  of:
\itemize{
 \item \code{composite_score}: An array with the calculated composite
 variable.
 \item \code{composite_metrics}: A matrix loadings and weights
 of the indicators.
 \item \code{composite_validity}: A matrix of composite reliability and
 validity metrics.
}
}
\description{
Calculate the composite score for the median family of weighting
  schemas.
}
\keyword{internal}
