% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_calc_mi_composite.R
\name{calc_mi_composite}
\alias{calc_mi_composite}
\title{Calculate Mutual-Information Composite Scores}
\usage{
calc_mi_composite(
  data,
  var,
  entropy = c("emp", "mm", "shrink", "sg"),
  nmi_method = c("geometric", "average"),
  threshold = 10,
  digits = 3,
  name = NULL,
  return_metrics
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{var}{A required vector of indicator column names.}

\item{entropy}{A string value reflecting the mutual information entropy
estimator from the \code{infotheo} package. Four estimators are available:
\code{emp} to compute the entropy of the empirical probability
distribution. Empirical entropy is suitable for simple calculations without
corrections. \code{mm} applies an asymptotic bias-corrected estimator
making it suitable for small sample sizes. \code{shrink} applies a
shrinkage estimate of the Dirichlet probability distribution to provide a
stable estimate useful for small sample sizes or sparse data. \code{sg}
applies a Schurmann-Grassberger estimate of the Dirichlet probability
distribution to serve as an alternative to the Shrinkage approach.}

\item{nmi_method}{A string value reflecting the method used for calculating
Normalized Mutual Information (NMI) values. \code{"average"} will normalize
MI values using the average entropies of variables A and B.
\code{"geometric"} will normalize MI values using the geometric mean of
entropies of variables A and B.}

\item{threshold}{An integer specifying the maximum number of unique values to
consider the input as discrete. Defaults to 10.}

\item{digits}{The decimal places for the metrics to be rounded to. Default is
3.}

\item{name}{A required string denoting the name of the composite variable.}

\item{return_metrics}{Logic to determine whether to return reliability and
validity metrics. Set to \code{TRUE} for a list of dataframes with
reliability and validity metrics.}
}
\value{
If \code{return_metrics = FALSE}, an array of the composite score is
  returned. If \code{return_metrics = TRUE}, a list is returned consisting
  of:
\itemize{
 \item \code{composite_score}: An array with the calculated composite
 variable.
 \item \code{composite_metrics}: A matrix loadings and weights
 of the indicators.
 \item \code{composite_validity}: A matrix of composite reliability and
 validity metrics.
}
}
\description{
Calculate the composite score for the mutual information family
  of weighting schemas.
}
\keyword{internal}
