% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_keep_higher_order_variable.R
\name{keep_higher_order_variable}
\alias{keep_higher_order_variable}
\title{Keep Higher Order Constructs from Composite List (Internal)}
\usage{
keep_higher_order_variable(composite_list, higher_order_variables)
}
\arguments{
\item{composite_list}{A required named list of vectors. Each name in the list
represents a composite variable, and the corresponding vector contains the
column names that are associated with the indicators comprising said
composite variable.}

\item{higher_order_variables}{A vector containing the names of higher order
variables.}
}
\value{
The function returns a new list that is a subset of
  \code{composite_list}, including only the vectors whose names are in
  \code{higher_order_variables.}
}
\description{
Takes a composite list as input and only keeps the vectors that
  match higher order constructs/variables.
}
\seealso{
[detect_higher_order_variable(), remove_higher_order_variable()]
}
\keyword{internal}
